Move graphviz-dot format writer to own module
